🐛 Debugging 1 | error fixed with adding double in ceil and now ans are right.

+++ Binary Search/MinEatingSpeed.java
@@ -36,7 +36,7 @@ public class MinEatingSpeed {
             
             for(int pile : piles){
                 // using ceil will round up the division
-                timeTaken += Math.ceil(pile / mid);
+                timeTaken += Math.ceil((double) pile / mid);
             } 
             System.out.println("[for] timeTaken: " + timeTaken);
